family-bot/requirements.txt
python-telegram-bot>=21.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
